refactor(endpoint-groups): move controller to separate file
